% Generated by roxygen2 (4.0.0): do not edit by hand
\name{BBfun}
\alias{BBfun}
\title{Numerical solver function for probability of a species existing at a site taking into
account other species and their competitive effects}
\usage{
BBfun(x, Pf, Alph, Nj, Nq)
}
\description{
Numerical solver function for probability of a species existing at a site taking into
account other species and their competitive effects
}

