% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ll.PLNT}
\alias{ll.PLNT}
\title{log likelihood function for Phylogenetic Latent Niche Traits Model (PLNT-Model)}
\usage{
ll.PLNT(comm, phylo, params)
}
\arguments{
\item{comm}{A community matrix with species in the columns and sites in the rows.
Matrix is composed of 0's and 1's where 0 is the absence of a species and 1 is the presence}

\item{phylo}{A phylogeny of the species in the community matrix in the form of the
Cholesky decomposition of the phylogenetic covariance matrix (can be generated a expectation
of Brownian motion evolution). Provide Cholesky decomposition so it does not have to be
performed during every iteration.}

\item{params}{}
}
\description{
log likelihood function for Phylogenetic Latent Niche Traits Model (PLNT-Model)
}

